question 1 and 2

+++ src/main/java/com/example/demo/repositories/MilitaryPostRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 
-public interface MilitaryPostRepo extends JpaRepository<MilitaryPost, Integer> {
+public interface MilitaryPostRepo extends JpaRepository<MilitaryPost, Long> {
 
 
 }
