+++ src/main/java/com/example/demo/repositories/OfficerRepo.java
@@ -3,7 +3,7 @@ package com.example.demo.repositories;
 import com.example.demo.models.Officer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OfficerRepo extends JpaRepository<Officer, Integer> {
+public interface OfficerRepo extends JpaRepository<Officer, Long> {
 
 
 }
